wsl-webcam-capture-example/requirements.txt
numpy
opencv-contrib-python
websocket-client
